27072017.1 Testes de calculo de descontos.

+++ src/test/java/br/com/huetech/calcdescontos/test/calculos/TestCalculosDescontosIT.java
@@ -34,7 +34,7 @@ public class TestCalculosDescontosIT extends BaseTestCase {
 		pageInicial.abaPaginaInicial();
 		pageInicial.irParaCalculoDescontos();
 		pageListagemProdutos.selecionaProdutoPorIdParaCalculoDesconto(Property.ID_PRODUTO);
-		int valorProduto = pageCalculaDescontoProduto.calcularDesconto(Property.TIPO_CLIENTE_A, Utils.geraNumeroEntre1_99());
+		Double valorProduto = pageCalculaDescontoProduto.calcularDesconto(Property.TIPO_CLIENTE_A, Utils.geraNumeroEntre1_99());
 		Log.msgFimTeste(nomeTeste);
 	}
 }
